fix ft_bzero

+++ ft_bzero.c
@@ -1,9 +1,7 @@
 
-void ft_bzero (void *s, int n) {
+void ft_bzero (void *s, size_t n) {
     unsigned char *toZero = s;
     while(n--){
-    
         *toZero++ = '\0';
     }
 }
-
